Revert "Refactor projects and time entries classes"
